### Group members:
- Bùi Huỳnh Long - 2352685 (Group CN02 - Double Deep Q NetWork )
- Nguyễn Hồng Huy - 2352392 (Group CN02 - FFD)
- Phạm Đăng Khải - 2352540 (Group CN02 - Column Generation)
- Nguyễn Thanh Tân - 2353079 (Group CN02 - Report)

### Algorithms:
### **1. Double Deep Q-Network (DDQN)**
Double Q-learning aims to improve the stability of standard Q-learning by reducing the overestimation bias.
This is done by using two networks: one for selecting actions (q_network) and another (target_network) for evaluating them.
Main Components:
Network Structure: A feedforward neural network with layers that use ReLU activations and batch normalization. It takes a state as input and outputs Q-values for each action.

Action Selection:

Exploration: With probability epsilon, the agent takes a random action (to explore). This is based on the available products and stocks, and the algorithm tries different placements and rotations.
Exploitation: With probability 1 - epsilon, the agent selects the action with the highest Q-value from the q_network.
Learning: The agent learns by comparing predicted Q-values with the target Q-values (which are computed using the Bellman equation). The target network is updated using a soft update to ensure stability.

Experience Replay: The agent stores its experiences (state, action, reward, next state, done) in a memory buffer. When enough experiences are collected, a random batch is sampled for learning.

### **2. First Fit Decreasing (FFD)**
FFD Policy: The algorithm sorts products based on their area (width × height), from largest to smallest, and then places them into the available stocks in the order they appear.
Steps:
Sorting: The products are sorted by their area in descending order.
Placement: The algorithm iterates over the sorted products and places them into the available stocks. For each stock, it checks if the product fits (either in its original or rotated orientation).
Best Fit: The algorithm checks if the product fits in the stock in either orientation (original or rotated), then selects the best fit based on the leftover area in the stock.
This method aims to minimize wasted space in the stock and ensures that products are placed efficiently.
### **3. Column Generation**
Column Generation Policy: This policy tries to place products based on column-based placements. It iterates over each product and tries to find the best placement within the available stocks using a combination of leftover area and usability.
Steps:
Placement Calculation: For each stock, the algorithm attempts to place the product (both in its original and rotated orientation).
Placement Scoring: A score is calculated for each potential placement, combining waste (unused area) and usability (how much of the stock's area can still be used after placing the product). The goal is to minimize the waste and maximize the usable area.
Best Placement: The algorithm chooses the placement with the lowest score and places the product in the corresponding stock.
Action Representation
Action Encoding: The action in the DDQN algorithm corresponds to a particular stock, product, rotation, and placement position. The action is represented as an index, which is decoded into a tuple of stock index, product index, placement position, and rotation.
State Representation: The state includes the characteristics of both the stocks and the products. The products are represented by their size, quantity, and orientation (rotation), while the stocks are represented by their size.
Experience Replay and Batch Learning
Memory Buffer: The agent stores past experiences in a buffer. After accumulating enough experiences, a batch of samples is used to train the network.
Target Network Update: The target network is updated with a soft update mechanism, where the weights of the target network are updated using a weighted average of the current and previous weights.

### November 1, 2024 - Project Kickoff
Activity: The team organizes a project kickoff meeting. Discuss project goals, scope, and initial task assignments. Set up development environment and required tools (Git, IDE, Python libraries).
Result: Git repository set up for the project, team members and roles defined, tools and libraries installed.

### November 10, 2024 - Theoretical Research and Method Selection
Activity: The team spends time researching existing algorithms for the problem, including Cutting Stock Problem algorithms. Choose the method to be used in the project, such as MRRF, Column Generation, Heuristics.
Result: Selection of the most appropriate method for the problem, preparation of theoretical documentation.

### November 15, 2024 - Development of Initial Model
Activity: Begin building the initial model for the cutting stock algorithm. Develop the simulation and test the input/output of the algorithm.
Result: The basic model is completed and can be run for testing.

### November 20, 2024 - Main Algorithm Implementation
Activity: Implement the main algorithm based on the researched theory. Program techniques like Lazy Initialization, Pattern Generation.
Result: The initial code of the algorithm is completed and can be run for testing.

### November 25, 2024 - Algorithm Analysis and Optimization
Activity: The team analyzes the implemented algorithms, identifying areas for optimization and improvement. Begin testing with different datasets to evaluate algorithm performance.
Result: Experimental datasets are implemented and results analyzed. Steps for improvement are proposed.

### November 30, 2024 - Test Suite and Documentation Creation
Activity: Build test cases and documentation for the report. Test edge cases and ensure the algorithm's accuracy.
Result: Test suite completed, initial report draft written.

### December 5, 2024 - Run Tests and Debug
Activity: Run tests with large datasets and check for errors. Debug issues encountered during testing.
Result: Code improved, minor bugs fixed, and performance optimized.

### December 10, 2024 - Algorithm Evaluation and Improvement
Activity: Analyze the results of the tests, compare the effectiveness of the applied algorithms, and select the optimal method. Implement improvements and optimize the source code.
Result: Improvements implemented, source code is more refined.

### December 12, 2024 - Optimization and Final Report Preparation
Activity: The team proceeds to optimize the algorithm to ensure accuracy and performance. Complete the theoretical section of the report and prepare the final documentation.
Result: The report is completed, covering all stages from theoretical research to experimental results.

### December 13, 2024 - Testing with Real-World Data
Activity: Run tests with real-world datasets from case studies. Collect data and analyze results.
Result: Real-world data collected and analyzed, results documented.

### December 14, 2024 - Finalization of Code and Report
Activity: Finalize the code, debug once more, and recheck any remaining errors. Complete the report and prepare the presentation.
Result: Code and report finalized, ready for project submission.