@huynhlong2706 2024-12-16
Em vừa sửa lại tiêu đề PR rồi ạ, thầy xem còn gì sai sót không ạ

Vào Th 2, 16 thg 12, 2024 lúc 18:42 Duc Quang Nguyen <
***@***.***> đã viết:

> Please revise the PR in Dec 16. Otherwise, your group will get 0 score for
> code.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/martinakaduc/MM241-Assignment/pull/146#issuecomment-2545394952>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/BBOETKLTAARFNYRS6NFJ6HD2F24AZAVCNFSM6AAAAABTTVXM62VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDKNBVGM4TIOJVGI>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>