@huynhlong2706 2024-12-16
Dạ thầy ơi cho em hỏi là em lỡ xoá PR cũ và cả Fork nên em vừa clone lại dự
án rồi push lên lại PR nhưng nó dính những file thầy đã merge nên em
xoá nên tình trạng bài em đang như trong ảnh ạ. Cho em hỏi là có được chấp
nhận không ạ ? Em cảm ơn thầy ạ.
[image: image.png]

Vào Th 2, 16 thg 12, 2024 vào lúc 18:42 Duc Quang Nguyen <
***@***.***> đã viết:

> Please revise the PR in Dec 16. Otherwise, your group will get 0 score for
> code.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/martinakaduc/MM241-Assignment/pull/146#issuecomment-2545394952>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/BBOETKLTAARFNYRS6NFJ6HD2F24AZAVCNFSM6AAAAABTTVXM62VHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDKNBVGM4TIOJVGI>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>